+++ android/src/com/hroo078/gxattack/Game/Objects/Enemies/UFOenemy.java
@@ -2,7 +2,7 @@ package com.hroo078.gxattack.Game.Objects.Enemies;
 
 public class UFOenemy extends Enemy {
 
-    public UFOenemy(int width, int height) {
+    public UFOenemy(float width, float height) {
         super(width, height);
         setTexture("enemyBlack1.png",true);
     }
